Impl WinLoseJudger; Fix DieSMB

+++ Assets/Scripts/SMB/Player/BattleBehaviour/DieSMB.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System;
 using UnityEngine;
+using Hashtable = ExitGames.Client.Photon.Hashtable;
 
 namespace Bunashibu.Kikan {
   public class DieSMB : StateMachineBehaviour {
@@ -12,11 +13,25 @@ namespace Bunashibu.Kikan {
       }
 
       _player.BodyCollider.enabled = false;
+
+      if (!_player.PhotonView.isMine)
+        return;
+
       SkillReference.Instance.DeleteAll();
 
       if (StageReference.Instance.StageData.Name == "Battle") {
         Action action = () => { _player.StateTransfer.TransitTo("Idle", animator); };
         _respawner.Respawn(action);
+      }
+
+      if (StageReference.Instance.StageData.Name == "FinalBattle") {
+        int[] alivePlayerCount = PhotonNetwork.room.CustomProperties["AlivePlayerCount"] as int[];
+        var list = MonoUtility.ToList<int>(alivePlayerCount);
+
+        list[_player.PlayerInfo.Team] -= 1;
+
+        var props = new Hashtable() {{ "AlivePlayerCount", list.ToArray() }};
+        PhotonNetwork.room.SetCustomProperties(props);
       }
     }
